Moving tests that do not use chex to _without_chex.py
